requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
Django
django-rest==0.8.7
djangorestframework==3.16.1
drf-yasg==1.21.14
fastapi==0.128.0
idna==3.11
inflection==0.5.1
packaging==26.0
psycopg2-binary==2.9.11
pydantic==2.11.10
pydantic_core==2.33.2
pytz==2025.2
PyYAML==6.0.3
six==1.17.0
sqlparse
starlette
typing-inspection
typing_extensions
tzdata
uritemplate
python-dotenv
